refactor: delete compute hash function in tools package
