+++ packages/tools/__tests__/computeHash.test.ts
@@ -1,7 +1,14 @@
 import { jest } from '@jest/globals';
-import { computeHash } from '../src/hash/computeHash';
+import { MultibaseEncodedSHA256 } from '../src/hash/multibase-encoded-256';
 
 describe('compute hash', () => {
+  let computeHash: any;
+
+  beforeEach(() => {
+    const plugin = new MultibaseEncodedSHA256();
+    computeHash = plugin.methods.computeHash;
+  });
+
   it('should compute hash', async () => {
     const hash = await computeHash('hello world');
     expect(hash).toBe('zQmaozNR7DZHQK1ZcU9p7QdrshMvXqWK6gpu5rmrkPdT3L4');
@@ -12,25 +19,18 @@ describe('compute hash', () => {
   });
 
   it('should throw error when value is not a string', async () => {
-    // @ts-expect-error
     await expect(computeHash(123)).rejects.toThrow('Value is invalid');
 
-    // @ts-expect-error
     await expect(computeHash(true)).rejects.toThrow('Value is invalid');
 
-    // @ts-expect-error
     await expect(computeHash(NaN)).rejects.toThrow('Value is invalid');
 
-    // @ts-expect-error
     await expect(computeHash(null)).rejects.toThrow('Value is invalid');
 
-    // @ts-expect-error
     await expect(computeHash(undefined)).rejects.toThrow('Value is invalid');
 
-    // @ts-expect-error
     await expect(computeHash({})).rejects.toThrow('Value is invalid');
 
-    // @ts-expect-error
     await expect(computeHash([])).rejects.toThrow('Value is invalid');
   });
 });
